Add FastScripts keyboard shortcuts via .fastscripts.plist
This was recommended by the author of FastScripts as a way to have
keyboard shortcuts work when switching between US and Dvorak.

This works if I quit and restart FastScripts when switching keyboard
layouts.

Here is the full message:

I think way back, over 15 years ago, when I first designed FastScripts,
I intentionally opted to associated keyboard shortcuts with the physical
key, thinking that it would be most likely that people would develop a
muscle memory for their keyboard shortcuts and thus want them to be the
same even when switching keyboard layouts. I also just confirmed that,
for example, even the system's keyboard shortcuts work as you describe,
associated by the letter/symbol rather than physical key.

In all this time I think you're the first who's run into a problem with
it, which is not to say it's not a problem, but I'm curious how much
people actually benefit from the system's and other apps' default
treatment. It seems like the FastScripts behavior would benefit you if
only this external Dvorak keyboard didn't need to masquerade as a US
layout keyboard in order to work properly?

I am going to think about this some more. It might be unfeasible for me
to upend the design of the app because I think the way I store the
keyboard shorcuts is strongly tied to the physical key-code, and it
would be tricky to even support as an option the "letter-based"
behavior.

I'm trying to think if there are any other possible solutions. I don't
officially support this yet, but FastScripts has a kind of secret,
undocumented feature that I added for somebody experimenting with using
the app with a bunch of users, whereby the script shortcuts can be
specified in the script folders themselves, as a special
".fastscripts.plist" file. For example such a file with the following
content:

{
    "ConfigurationFileVersion" = 1;
    KeyboardShortcuts = (
        {
            RelativePath = "Open Order.scpt";
            ShortcutModifiers = 3277066;
            ShortcutKeyCode = 83;
        },
        {
            RelativePath = "CRM Shortcuts.scpt";
            Shortcut = "Command-Control-Shift-Option-C";
        },
    );
}

Would create two keyboard shortcuts for the scripts at the relative
path. Notice how the second form uses a "Shortcut" value that is based
on a character description? You might want to play around with this if
you think it would be helpful! Unfortunately it's not possible to
represent every key reliably with this notation. Also, if you adopt this
as a workaround, you might need to quit and relaunch FastScripts when
you switch keyboards, to get it to reinitialize the shortcuts based on
interpretation of those key specifiers.