Remove .fastscripts.plist
For now, I'm removing the plist file, with a plan to just live with the
fact that I need to do key combinations based on physical keys on my
laptop when in dvorak.

My main reason for this is that I think it will be a pain to restart
fastscripts whenever I change layouts, and even more of a pain if I
forget to do this sometimes and so do the wrong key combo.